Created Comments service, its working implementation, the web controller is also created and fully working with get, post methoods

+++ src/main/java/com/cg/repo/FriendshipRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cg.entity.Friendship;
 
-public interface FriendshipRepo extends JpaRepository<Friendship, Integer>{
+public interface FriendshipRepo extends JpaRepository<Friendship, Long>{
 
 }
